+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.1] - 2023-07-10
+
+### Fixed
+
+- Fixed a residual bug in `calibrate.bsh` from renaming row/col to x/y.
+
 ## [1.1.0] - 2023-07-10
 
 ### Added
@@ -68,7 +74,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added a `calibrate.bsh` script to run the Fourier ptychography calibration routine.
 
-[Unreleased]: https://github.com/leb-epfl/mr-freeze/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/leb-epfl/mr-freeze/compare/v1.1.1...HEAD
+[1.1.1]: https://github.com/leb-epfl/mr-freeze/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/leb-epfl/mr-freeze/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/leb-epfl/mr-freeze/releases/tag/v1.0.0
 [leb-freeze-0.0.0]: https://github.com/leb-epfl/mr-freeze/releases/tag/leb-freeze-v0.0.0